initial error handler and usage
