+++ client_src/js/services/octofarm_client.service.js
@@ -25,8 +25,9 @@ axios.interceptors.response.use(
   function (error) {
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
-    console.error(JSON.stringify(error, null, 4));
-    throw new ApplicationError(HTTPError.RESOURCE_NOT_FOUND);
+    if (error.response.status === 404) {
+      throw new ApplicationError(HTTPError.RESOURCE_NOT_FOUND);
+    }
   }
 );
