+++ views/partials/octofarm-error.modal.ejs
@@ -7,11 +7,12 @@
 
                 <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
             </div>
-            <div class="modal-body text-center">
+            <div class="modal-body text-center pt-0">
                 <div id="apiErrorMessage"></div>
-                <p>
-                    Double check your logs and refresh to continue.
-                </p>
+                <div class="py-3">
+                    If this issue persists, please submit a bug report with the below information...
+                </div>
+                <div id="apiDeveloperInfo"></div>
             </div>
         </div>
     </div>
